change package name from example to jobandotrucking
