+++ app/src/androidTest/java/com/jobandotrucking/eld/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.eld;
+package com.jobandotrucking.eld;
 
 import android.content.Context;
