+++ app/src/main/java/com/jobandotrucking/eld/DBHelper.java
@@ -1,4 +1,4 @@
-package com.example.eld;
+package com.jobandotrucking.eld;
 
 import android.content.ContentValues;
 import android.content.Context;
