+++ app/src/main/java/com/jobandotrucking/eld/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.eld;
+package com.jobandotrucking.eld;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.eld.R;
 
 public class MainActivity extends AppCompatActivity {
