+++ app/src/main/java/com/jobandotrucking/eld/contactUsFragment.java
@@ -1,4 +1,4 @@
-package com.example.eld;
+package com.jobandotrucking.eld;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,12 +7,14 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
+import com.example.eld.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link workingHoursFragment#newInstance} factory method to
+ * Use the {@link contactUsFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class workingHoursFragment extends Fragment {
+public class contactUsFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -23,7 +25,7 @@ public class workingHoursFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public workingHoursFragment() {
+    public contactUsFragment() {
         // Required empty public constructor
     }
 
@@ -33,11 +35,11 @@ public class workingHoursFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment workingHoursFragment.
+     * @return A new instance of fragment ContactUsFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static workingHoursFragment newInstance(String param1, String param2) {
-        workingHoursFragment fragment = new workingHoursFragment();
+    public static contactUsFragment newInstance(String param1, String param2) {
+        contactUsFragment fragment = new contactUsFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -58,6 +60,6 @@ public class workingHoursFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_working_hours, container, false);
+        return inflater.inflate(R.layout.fragment_contact_us, container, false);
     }
 }
