+++ app/src/main/res/layout/activity_add_new_driver.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".addNewDriver">
+        tools:context="com.jobandotrucking.eld.addNewDriver">
 
     <TextView
         android:id="@+id/textView"
