+++ app/src/main/res/layout/activity_admin_dashboard.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_weight="1"
-    tools:context=".adminDashboard">
+    tools:context="com.jobandotrucking.eld.adminDashboard">
 
     <TextView
         android:id="@+id/textView"
