+++ app/src/main/res/layout/activity_assign_job.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".assignJob">
+        tools:context="com.jobandotrucking.eld.assignJob">
 
         <TextView
             android:id="@+id/textView"
