+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bgmain"
-    tools:context=".MainActivity">
+    tools:context="com.jobandotrucking.eld.MainActivity">
 
     <EditText
         android:id="@+id/uname"
