+++ app/src/main/res/layout/fragment_about_us.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".aboutUsFragment">
+    tools:context="com.jobandotrucking.eld.aboutUsFragment">
 
     <TextView
         android:layout_width="wrap_content"
