+++ app/src/main/res/layout/fragment_contact_us.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".contactUsFragment">
+    tools:context="com.jobandotrucking.eld.contactUsFragment">
 
     <TextView
         android:layout_width="wrap_content"
