+++ app/src/main/res/layout/fragment_maps_direction.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mapsDirectionFragment">
+    tools:context="com.jobandotrucking.eld.mapsDirectionFragment">
 
     <fragment
         android:id="@+id/googleMap"
