+++ app/src/main/res/layout/fragment_past_records.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".pastRecordsFragment">
+    tools:context="com.jobandotrucking.eld.pastRecordsFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
