+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".settingsFragment">
+    tools:context="com.jobandotrucking.eld.settingsFragment">
 
     <TextView
         android:layout_width="wrap_content"
