+++ app/src/main/res/layout/fragment_working_hours.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".workingHoursFragment">
+    tools:context="com.jobandotrucking.eld.workingHoursFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
