+++ app/src/test/java/com/jobandotrucking/eld/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.eld;
+package com.jobandotrucking.eld;
 
 import org.junit.Test;
